Sort to worker + debounce filter
