+++ src/common/util/render-status.ts
@@ -1,3 +1,9 @@
 export const afterNextRender = (cb: () => void): void => {
   requestAnimationFrame(() => setTimeout(cb, 0));
 };
+
+export const nextRender = () => {
+  return new Promise((resolve) => {
+    afterNextRender(resolve);
+  });
+};
