+++ src/components/data-table/sort_filter_worker.ts
@@ -1,0 +1,63 @@
+import {
+  DataTabelColumnContainer,
+  DataTabelColumnData,
+  DataTabelRowData,
+  SortingDirection,
+} from "./ha-data-table";
+
+export const filterData = (
+  data: DataTabelRowData[],
+  columns: DataTabelColumnContainer,
+  filter: string
+) =>
+  data.filter((row) => {
+    return Object.entries(columns).some((columnEntry) => {
+      const [key, column] = columnEntry;
+      if (column.filterable) {
+        if (
+          (column.filterKey ? row[key][column.filterKey] : row[key])
+            .toUpperCase()
+            .includes(filter)
+        ) {
+          return true;
+        }
+      }
+      return false;
+    });
+  });
+
+export const sortData = (
+  data: DataTabelRowData[],
+  column: DataTabelColumnData,
+  direction: SortingDirection,
+  sortColumn: string
+) =>
+  data.sort((a, b) => {
+    let sort = 1;
+    if (direction === "desc") {
+      sort = -1;
+    }
+
+    let valA = column.filterKey
+      ? a[sortColumn][column.filterKey]
+      : a[sortColumn];
+
+    let valB = column.filterKey
+      ? b[sortColumn][column.filterKey]
+      : b[sortColumn];
+
+    if (typeof valA === "string") {
+      valA = valA.toUpperCase();
+    }
+    if (typeof valB === "string") {
+      valB = valB.toUpperCase();
+    }
+
+    if (valA < valB) {
+      return sort * -1;
+    }
+    if (valA > valB) {
+      return sort * 1;
+    }
+    return 0;
+  });
